requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cmake==3.17.3
cos-python-sdk-v5==1.9.2
crcmod==1.7
dicttoxml==1.7.4
Django==1.11.28
django-redis==4.11.0
dlib==19.21.1
idna==2.10
jieba==0.42.1
joblib==1.0.1
numpy==1.18.5
pandas==1.2.2
Pillow==6.2.1
pycryptodome==3.9.9
python-dateutil==2.8.1
pytz==2020.5
qcloud-python-sts==3.0.5
qcloudsms-py==0.1.4
redis==3.4.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.2
xlwt==1.3.0
